+++ ModLoader/ModLoader.cs
@@ -28,6 +28,8 @@ namespace Sheepy.Modnix {
       internal readonly static Version PPML_COMPAT = new Version( 0, 2 );
 
       public static string ModDirectory { get; private set; }
+      public static string LoaderPath => Assembly.GetExecutingAssembly().Location;
+      public static string DnFrameworkDir => Path.Combine( Environment.GetFolderPath( Environment.SpecialFolder.Windows ), "Microsoft.NET/Framework/v4.0.30319" );
 
       #region Initialisation
       private static bool RunMainPhaseOnInit;
@@ -97,8 +99,8 @@ namespace Sheepy.Modnix {
             LogGameVersion();
          }
          var corlib = new Uri( typeof( string ).Assembly.CodeBase ).LocalPath;
-         Log.Verbo( ".Net/{0}; mscorlib/{1} {2}", Environment.Version, FileVersionInfo.GetVersionInfo( corlib ).FileVersion, corlib );
          LoadSettings();
+         Log.Verbo( ".Net/{0}; mscorlib/{1} {2}", Environment.Version, FileVersionInfo.GetVersionInfo( corlib ).FileVersion, corlib );
       } } catch ( Exception ex ) { Log?.Error( ex ); } }
 
       internal static volatile Assembly PpmlAssembly;
@@ -106,17 +108,21 @@ namespace Sheepy.Modnix {
       // Dynamically load embedded dll
       private static Assembly ModLoaderResolve  ( object domain, ResolveEventArgs dll ) { try {
          var name = dll.Name;
-         Log.Trace( "Resolving {0}", name );
          var app = domain as AppDomain ?? AppDomain.CurrentDomain;
-         if ( name.StartsWith( "PhoenixPointModLoader,", StringComparison.OrdinalIgnoreCase ) )
+         if ( name.StartsWith( "PhoenixPointModLoader,", StringComparison.OrdinalIgnoreCase ) ) {
+            Log.Info( "Loading embedded PPML" );
             return PpmlAssembly = app.Load( GetResourceBytes( "PPML_0_2.dll" ) );
+         }
          if ( name.StartsWith( "System." ) && dll.Name.Contains( ',' ) ) { // Generic system library lookup
             var file = dll.Name.Substring( 0, dll.Name.IndexOf( ',' ) ) + ".dll";
-            var target = Path.Combine( Environment.GetFolderPath( Environment.SpecialFolder.Windows ), "Microsoft.NET/Framework/v4.0.30319", file );
+            var target = Path.Combine( DnFrameworkDir, file );
             if ( File.Exists( target ) ) {
                Log.Info( "Loading {0}", target );
                return Assembly.LoadFrom( target );
             }
+         } else {
+            Log.Warn( "Cannot resolve {0}", name );
+            Log.Flush(); // The app may crash right after the failure, so flush the log now
          }
          return null;
       } catch ( Exception ex ) { Log?.Error( ex ); return null; } }