+++ MainGUI/Resources/Changelog.md
@@ -8,6 +8,7 @@ Modnix Changelog
 * Mod: Loader log now tries to capture game crashs originating from the scope of CLR.
 * Mod: Loader log now warns on unresolved assemblies.
 * Mod: Loader now logs mscorlib version and end of phases at verbose level.
+* Mod: Requires and Disables are now resolved together in each resolve loop, instead of only Requires.
 * Mod: Flags and Actions in mod_info (intended for Modnix 3) now leave a notice in Manager and log.
 
 # Version 2.3.1, 2020-05-03